// src/app/models/issue.model.ts
export class IssueModel {
    _id?: string = '';
    name: string = '';
    issue: string = '';
    issueJonour: string = '';
    place: string = '';
    pincode: string = '';
    complexity: string = '';
    status: string = ''; 
    phone:string = '';
    email:string = '';
    city:string = '';
    state:string = '';
    address:string = '';
    gender:string = '';
 }